refactor: MineBoard 생성 로직을 MineBoardBuilder 로 위임
