+++ src/main/kotlin/minesweeper/domain/board/MineBoardBuilder.kt
@@ -1,6 +1,15 @@
 package minesweeper.domain.board
 
+import minesweeper.domain.cell.Cells
+import minesweeper.domain.cell.Empty
+import minesweeper.domain.cell.Mine
+import minesweeper.domain.common.NumberSet
+import minesweeper.domain.common.Position
 import minesweeper.domain.common.PositiveInt
+import minesweeper.domain.common.contains
+import minesweeper.domain.common.div
+import minesweeper.domain.common.rangeTo
+import minesweeper.domain.common.rem
 import kotlin.properties.Delegates
 
 class MineBoardBuilder {
@@ -25,5 +34,30 @@ class MineBoardBuilder {
         mineStrategy = strategy
     }
 
-    fun build() = MineBoard(width, height, numberOfMines, mineStrategy)
+    fun build(): MineBoard {
+        val size = width * height
+
+        require(numberOfMines in (0..size)) { "number of mines must be within range of 0 ~ $size" }
+
+        val mineIndices = NumberSet.of(mineStrategy(size, numberOfMines))
+        val board = createBoard(width, height, mineIndices)
+
+        return MineBoard.of(mineIndices, board)
+    }
+
+    private fun createBoard(width: PositiveInt, height: PositiveInt, mineIndices: NumberSet): Board {
+        val size = (width * height).value
+        val cells = Cells(
+            List(size) {
+                val x = it % width
+                val y = it / width
+                if (it in mineIndices) {
+                    Mine(Position.of(x, y))
+                } else {
+                    Empty(Position.of(x, y))
+                }
+            }
+        )
+        return Board.of(width, height, cells)
+    }
 }
